+++ src/output/render/render-json.ts
@@ -13,7 +13,8 @@ export const renderViolationsJsonObject = (
   return pairs.flatMap(([_, violations]) => {
     const first = violations[0];
     const { fileName } = first;
-    const rawLocation = first.field?.location ?? first.model.location;
+    const rawLocation =
+      first.field?.location ?? first.model?.location ?? first.enum?.location;
     if (!rawLocation) {
       throw new Error('No location');
     }