+++ src/rule-definitions.ts
@@ -1,4 +1,6 @@
 import type { RuleDefinition } from '#src/common/rule.js';
+import enumNamePascalCase from '#src/rules/enum-name-pascal-case.js';
+import enumValueSnakeCase from '#src/rules/enum-value-snake-case.js';
 import fieldNameMappingSnakeCase from '#src/rules/field-name-mapping-snake-case.js';
 import fieldOrder from '#src/rules/field-order.js';
 import forbidField from '#src/rules/forbid-field.js';
@@ -12,6 +14,8 @@ import requireFieldType from '#src/rules/require-field-type.js';
 import requireField from '#src/rules/require-field.js';
 
 export default [
+  enumNamePascalCase,
+  enumValueSnakeCase,
   fieldNameMappingSnakeCase,
   fieldOrder,
   forbidField,